requirements.txt
pytest
python-dateutil
ics
